% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{as_database}
\alias{as_database}
\title{Helper function for converting a RMSFasta object to a table}
\usage{
as_database(x)
}
\arguments{
\item{x}{rmfasta data object}

\item{...}{unused legacy}
}
\value{
a data.frame
}
\description{
Helper function for converting a RMSFasta object to a table
}
