% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{A helper function to get the string defined by the regex}
\usage{
extract(string = NULL, regex = NULL)
}
\arguments{
\item{string}{A section of text}

\item{regex}{A list of regular expressions defining bits of text to exract}
}
\description{
\code{extract()} get the current string based on regex search
}
