% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{make_table}
\alias{make_table}
\title{Helper function for converting a msfastar object to a table}
\usage{
make_table(x, .drop = NULL)
}
\arguments{
\item{x}{An rmfasta data object}

\item{.drop}{Call out columns to drop from the data.frame}
}
\description{
Helper function for converting a msfastar object to a table
}
