% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{msfastar-package}
\alias{msfastar-package}
\title{msfastar: A light-weight FASTA reader and parser}
\description{
A light-weight R package for parsing FASTA files into R usable formats and generating expected peptides from a given regex.
}
\author{
\strong{Maintainer}: Jeff Jones \email{jeffj@caltech.edu} (\href{https://orcid.org/0000-0002-7142-2222}{ORCID})

}
