% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptides.R
\name{peptides}
\alias{peptides}
\title{Proteolytic digest a sequence}
\usage{
peptides(
  sequence = NULL,
  enzyme = "[KR]",
  partial = 0:3,
  length = c(6, 30),
  remove_nterm_m = TRUE
)
}
\arguments{
\item{sequence}{as character string}

\item{enzyme}{a character string regular expression use to proteolytically digest
the sequence.
\itemize{
\item \verb{[KR]} ... trypsin
\item \verb{[KR](?!P)} ... trypsin not at P
\item \verb{[R](?!P)} ... arg-c
\item \verb{[K](?!P)} ... lys-c
\item \verb{[FYWL](?!P)} ... chymotrypsin
\item \verb{[BD]} ... asp-n
\item \verb{[D]} ... formic acid
\item \verb{[FL]} ... pepsin-a
}}

\item{partial}{a numeric representing the number of incomplete enzymatic sites (mis-clevage).}

\item{length}{as numeric vactor representing the minimum and maximum sequence lengths.}
}
\value{
a vector
}
\description{
\code{peptides()} Generates peptide sequences based on \code{enzyme} and \code{partial} inputs.
}
\examples{
library(msfastar)
sequence <- "SAMERSMALLKPSAMPLERSEQUENCE"
peptides(sequence)

peptides(sequence, enzyme = "[L]", partial = 2, length = c(1,12))

}
