% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{regex}
\alias{regex}
\title{Get/Set the FASTA meta data regex}
\usage{
regex(params = NULL)
}
\arguments{
\item{params}{as list}
}
\value{
a list
}
\description{
\code{regex()} gets and sets the current regex patters to assist the \code{read_fasta()} function.
This simply provides the structure needed to read_fasta the fasta file, a custom list
can also be supplied. To set elements in the \code{regex()} function, simply provide a
list with complementary names to over-write the current list.
}
\examples{
library(msfastar)
regex()

regex(list("accession" = "sp\\\\|[A-Z]"))

}
