% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fasta.R
\name{save_fasta}
\alias{save_fasta}
\title{The main function for parsing a fasta file}
\usage{
save_fasta(fasta = NULL, path = "test.fasta")
}
\arguments{
\item{fasta_object}{a fasta db object}
}
\value{
a success TRUE|FAIL
}
\description{
\code{save_fasta()} save the file
}
