% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.msfastar}
\alias{subset.msfastar}
\title{Create a data subset}
\usage{
\method{subset}{msfastar}(x = NULL, ..., .verbose = TRUE)
}
\arguments{
\item{x}{An ms2spectra data object}

\item{...}{A three part expression (eg. x == a)}

\item{.verbose}{A boolean to print messages}
}
\description{
\code{subset()} is the main function for sub-setting spectra data from a ms2spectra
data-object based on a regular expression and targeted annotation. This function
will return a smaller ms2spectra data-object.
}
