% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write contents to a fasta file}
\usage{
write_fasta(fasta = NULL, path = NULL)
}
\arguments{
\item{fasta}{A fasta db object}

\item{path}{A character string of the path to the fasta formatted file}
}
\description{
\code{write_fasta()} is the main function for writing a msfastar data object to a
fasta file. This function is currently in BETA, it will just print the contents
to stdout.
}
