% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{as.data.frame.msfastar}
\alias{as.data.frame.msfastar}
\title{Helper function for converting a msfastar object to a table}
\usage{
\method{as.data.frame}{msfastar}(x, .drop = NULL, ...)
}
\arguments{
\item{x}{rmfasta data object}

\item{...}{unused legacy}
}
\value{
a data.frame
}
\description{
Helper function for converting a msfastar object to a table
}
