% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{as_database}
\alias{as_database}
\title{Helper function for converting a msfastar object to a table}
\usage{
as_database(x)
}
\arguments{
\item{x}{An rmfasta data object}
}
\description{
Helper function for converting a msfastar object to a table
}
