% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Proteolytic digest a read_fastad fasta list}
\usage{
digest(x = NULL, ...)
}
\arguments{
\item{x}{an rmsfasta data object}

\item{...}{parameters for \code{peptides()}}
}
\value{
a rmsfasta data object
}
\description{
\code{digest()} Generates peptide sequences based on \emph{enzyme} and \emph{partial} inputs.
Only works with the "list" output of the \code{read_fasta()} function
}
\examples{
library(rmsfasta)
proteins <- read_fasta("~/Local/fasta/ecoli_UniProt.fasta")

proteins <- digest(proteins, enzyme = "[K]", partial = 2)

}
