% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Get the string defined by the regex}
\usage{
extract(string = NULL, regex = NULL)
}
\arguments{
\item{string}{a character}

\item{regex}{a list}
}
\value{
a list
}
\description{
\code{extract()} get the current string based on regex
}
