% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{get_peptides}
\alias{get_peptides}
\title{Get all peptides from msfastar object as a vector}
\usage{
get_peptides(x = NULL)
}
\arguments{
\item{x}{An rmfasta data object}
}
\description{
\code{get_peptides()} will return all peptides as single string vector
}
