% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe.R
\name{make_table}
\alias{make_table}
\title{Helper function for converting a RMSFasta object to a table}
\usage{
make_table(x, .drop = NULL)
}
\arguments{
\item{x}{rmfasta data object}
}
\value{
a data.frame
}
\description{
Helper function for converting a RMSFasta object to a table
}
