% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.msfastar}
\alias{print.msfastar}
\title{A helper function for the print definition of a msfastar object}
\usage{
\method{print}{msfastar}(x, ...)
}
\arguments{
\item{x}{An rmfasta data object}

\item{...}{Unused legacy}
}
\description{
A helper function for the print definition of a msfastar object
}
