% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{regex}
\alias{regex}
\title{Get/Set the FASTA meta data regex}
\usage{
regex(params = NULL)
}
\arguments{
\item{params}{A list of the meta data to be extracted from the fasta file

Example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  "accession" = "(?<=|)[A-Z0-9]\{5,12\}",
  "protein_name" = "(?<=|)[A-Z0-9\\\\_]\{8,\}(?=\\\\s)",
  "gene_name" = "(?<=GN\\\\=).*?(?=\\\\s..\\\\=)",
  "organism" = "(?<=OS\\\\=).*?(?=\\\\s..\\\\=)",
  "description" = "(?<=\\\\s).*?(?=\\\\s..\\\\=)",
  "sequence" = "[A-Z]"
)
}\if{html}{\out{</div>}}}
}
\description{
\code{regex()} gets and sets the current regex patters to assist the \code{read_fasta()} function.
This simply provides the structure needed to read_fasta the fasta file, a custom list
can also be supplied. To set elements in the \code{regex()} function, simply provide a
list with complementary names to over-write the current list.
}
\examples{
library(msfastar)
regex()

regex(list("accession" = "sp\\\\|[A-Z]"))

}
